decouple Jehanne from 9P2000 (as much as needed)
Jehanne is going to use a new file protocol, but Plan 9 is really
coupled with 9P2000.

Renamed fcall.h as 9P2000.h and introduced specific constants such
as NP_OREAD, NP_OWRITE and so on, so that we can use different values
in the kernel and new protocol.

Renamed devmnt to devninep, since it's actually a device serving 9P2000
file systems.

Also, fixed 9P2000 support in Jehanne, that was broken with the introduction
of OSTAT.

+++ sys/src/kern/386/devether.c
@@ -56,13 +56,13 @@ etherstat(Chan* chan, uint8_t* dp, long n)
 }
 
 static Chan*
-etheropen(Chan* chan, int omode)
+etheropen(Chan* chan, unsigned long omode)
 {
 	return netifopen(&etherxx[chan->devno]->netif, chan, omode);
 }
 
 static Chan*
-ethercreate(Chan* _1, char* _2, int _3, int _4)
+ethercreate(Chan* _1, char* _2, unsigned long _3, unsigned long _4)
 {
 	error(Eperm);
 	return nil;
