+++ sys/src/kern/amd64/devvga.c
@@ -124,9 +124,9 @@ vgareset(void)
 {
 	/* reserve the 'standard' vga registers */
 	if(ioalloc(0x2b0, 0x2df-0x2b0+1, 0, "vga") < 0)
-		panic("vga ports already allocated"); 
+		panic("vga ports already allocated");
 	if(ioalloc(0x3c0, 0x3da-0x3c0+1, 0, "vga") < 0)
-		panic("vga ports already allocated"); 
+		panic("vga ports already allocated");
 	addarchfile("realmodemem", 0660, rmemread, rmemwrite);
 }
 
@@ -151,7 +151,7 @@ vgastat(Chan* c, uint8_t* dp, long n)
 }
 
 static Chan*
-vgaopen(Chan* c, int omode)
+vgaopen(Chan* c, unsigned long omode)
 {
 	VGAscr *scr;
 	static char *openctl = "openctl\n";
@@ -160,7 +160,7 @@ vgaopen(Chan* c, int omode)
 	if ((uint32_t)c->qid.path == Qvgaovlctl) {
 		if (scr->dev && scr->dev->ovlctl)
 			scr->dev->ovlctl(scr, c, openctl, strlen(openctl));
-		else 
+		else
 			error(Enonexist);
 	}
 	return devopen(c, omode, vgadir, nelem(vgadir), devgen);
@@ -230,7 +230,7 @@ vgaread(Chan* c, void* a, long n, int64_t off)
 				scr->gscreen->r.max.x, scr->gscreen->r.max.y,
 				scr->gscreen->depth, chantostr(chbuf, scr->gscreen->chan));
 
-			if(Dx(scr->gscreen->r) != Dx(physgscreenr) 
+			if(Dx(scr->gscreen->r) != Dx(physgscreenr)
 			|| Dy(scr->gscreen->r) != Dy(physgscreenr))
 				len += snprint(p+len, READSTR-len, "actualsize %dx%d\n",
 					physgscreenr.max.x, physgscreenr.max.y);
@@ -381,7 +381,7 @@ vgactl(Cmdbuf *cb)
 		physgscreenr = Rect(0,0,x,y);
 		scr->gscreen->clipr = physgscreenr;
 		return;
-	
+
 	case CMpalettedepth:
 		x = strtoul(cb->f[1], &p, 0);
 		if(x != 8 && x != 6)
@@ -396,7 +396,7 @@ vgactl(Cmdbuf *cb)
 		if(scr->dev && scr->dev->drawinit)
 			scr->dev->drawinit(scr);
 		return;
-	
+
 	case CMlinear:
 		if(cb->nf!=2 && cb->nf!=3)
 			error(Ebadarg);
@@ -408,7 +408,7 @@ vgactl(Cmdbuf *cb)
 		if(screenaperture(size, align) < 0)
 			error("not enough free address space");
 		return;
-/*	
+/*
 	case CMmemset:
 		memset((void*)strtoul(cb->f[1], 0, 0), atoi(cb->f[2]), atoi(cb->f[3]));
 		return;
@@ -417,11 +417,11 @@ vgactl(Cmdbuf *cb)
 	case CMblank:
 		drawblankscreen(1);
 		return;
-	
+
 	case CMunblank:
 		drawblankscreen(0);
 		return;
-	
+
 	case CMblanktime:
 		blanktime = strtoul(cb->f[1], 0, 0);
 		return;
@@ -450,7 +450,7 @@ vgactl(Cmdbuf *cb)
 		else
 			break;
 		return;
-	
+
 	case CMhwblank:
 		if(strcmp(cb->f[1], "on") == 0)
 			hwblank = 1;
