+++ sys/src/kern/port/devproc.c
@@ -294,7 +294,7 @@ nonone(Proc *p)
 }
 
 static Chan*
-procopen(Chan *c, int omode)
+procopen(Chan *c, unsigned long omode)
 {
 	Proc *p;
 	Pgrp *pg;
